lawn: pass a count into read channels
We'd like to be able to pass a count of the number of bytes into the
`read` function to indicate how much data is to be read.  So far, this
hasn't been a problem, but it isn't always the case that a caller will
be willing to read large amounts of data.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>