Convert client channel operations to an async reader/writer
Currently, we attempt to poll all the selectors of the remote channel,
and if they return that they're ready, we process all of the channels.
However, this leads to a problem that we must poll in order to read or
write, even if the remote side is constantly ready.

Instead, let's adjust our code to use implementations of the `AsyncRead`
and `AsyncWrite` traits that simply read or write as much as possible
until they return `EAGAIN`, only at which point do they poll.  This
greatly simplifies our main loop, since all of the logic sits inside the
read and write structs.

This may substantially improve performance of channels like SFTP because
there's no longer any need to perform a needless round trip on every
read and write.

As a result, though, the requirements on our reader and writer types
must change to include `'static`.  Thus, we can no longer use mutable
references here, so change our proxy code to avoid doing that by passing
`self` to `run_server` instead of `&mut self`.  This also removes the
need for the proxy to be mutable in most cases, so remove that as well.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>