Improve efficiency of parsing packets
Right now, it can take 20 ms to parse a packet, in part because we do so
inefficiently.  We read the entire packet into one buffer, and then
split it out, copying the body into a separate packet.  This is not very
efficient, and we can do better.

Instead, copy the header data into a single block and then read the body
data directly into a `BytesMut`, which we can turn into a `Bytes`
cheaply.  As a result, parsing this data now takes only 8 to 10 ms.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>