lawn-protocol: box messages
When we read this message, we're going to send it through a channel in
the server.  That means that we'll incur a copy which we could easily
avoid.  Instead, let's box this value, so we're sending a pointer
through the channel instead.  We'll reuse this same pointer when
processing the message, since we borrow the value.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>