Use trace! macro to improve performance
When we include trace data, we typically need to format a string.  If
we're not actually tracing data, that's an expense we'd like to avoid.

In several places in our code, we called `logger.trace` manually,
including a formatting string, and, as a result, we format the data even
when we're not tracing.  Switch to the `trace!` macro, which checks the
log level first to avoid calling the formatter if it's not needed.

This can improve performance when piping large amounts of data to a
command from 104–109 MB/s to 108–117 MB/s.