lawn-protocol: avoid needless copies when serializing
Right now, we end up making a lot of copies in our serialization code.
If we take an already-serialized message, we copy it into a buffer,
which we don't pre-allocate in any way.  If it needs to be serialized,
we serialize the message, and then copy it just as if it were an
already-serialized message.  There's even more copies in the handler
code as well.  All of this is bad for performance.

Instead, try to avoid copying existing messages by simply serializing a
header and then writing the header and message without making any
copies.  If we do need to serialize a structure, do that by first
writing the entire header with a zero size (which is never valid), plus
the other fields (all with pre-allocation), serializing the body after
it, and then re-writing the size field.  This avoids as many allocations
and copies as possible.

All of this has a substantial impact on performance.  Performing a git
clone of git/git over HTTPS, making a local copy, and deleting it all
over SFTP with sshfs goes from taking 63 seconds to 53.  In addition,
piping 1 GiB of zeros to `sha256sum` with `lawn run` goes from 115 MB/s
to 297 MB/s, also very impressive.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>