client: use Arc<Connection> where possible
When processing channels, we want to pass some of our data across thread
boundaries.  However, with a regular `&self`, we can't do that, since we
can't be certain that we'll live long enough.  Instead, let's use an
`Arc` for `self`, so that we can pass it around across thread boundaries
and be guaranteed that it will always live long enough.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>