Make our client I/O Send and 'static
In the future, we're going to send the I/O objects into a separate Tokio
task, so let's make sure that these values are `Send` and `'static` so
we can do so.

Note that this means that our Proxy needs to become a single-shot
operation, which is completely fine, and thus the variable that holds it
no longer needs to be `mut`.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>