Improve channel performance
Currently, we attempt to poll all the selectors of the remote channel,
and if they return that they're ready, we process all of the channels.
However, this leads to a problem that we must poll in order to read or
write, even if the remote side is constantly ready.

Instead, let's run three separate tasks to process each channel as
frequently as possible and only poll the individual channel if it times
out.  This lets us maximize throughput and performance, since for
channels like SFTP, there's no need for an expensive round trip on every
read and write.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>