Makefile: add a target to build the changelog for a version
In the future, we'll want to build the changelog into a Markdown file
for the purpose of uploading it as a GitHub release.  Let's add a target
to the Makefile to do this for us in a new directory called `assets`.

Since this directory contains autogenerated assets, let's ignore it and
clean it up when we're done.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>