lawn: add serialization tests
Since we've publicly defined the protocol to use in terms of CBOR, let's
define some serialization tests so that we can verify our data
serializes as we expect and document.  Implement some basic tests that
do encoding and decoding of various types so that we can have confidence
in our serialization.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>