Add timeouts for `lawn proxy`
The SSH agent protocol operates such that the client always speaks
first.  Normally, this is true of the Lawn protocol as well, but we can
receive exit notifications from the server when running a command, so we
also need a way for the server to send asynchronous responses.

The solution, unfortunately, is to poll.  We have traditionally had a
100 ms delay after which we send an empty poll packet, such that we can
receive asynchronous responses no later than this.

However, as this stands, we have a slight performance problem.  If the
server response is not completely ready when we read it, we send back an
empty response and we need to wait for the next poll.  Instead, let's
wait a few milliseconds to see if a response is ready, and if so, send
it right away.

This actually dramatically improves performance: a `lawn query
test-connection` with `lawn proxy` on the same machine goes from 0.410 s
to 0.011 s, which is a substantial improvement.  To see if we can
improve performance a little more, adjust the default timeout for
polling to 50 ms.

Because sometimes these values won't work in some case or another, let's
also make them configurable so that users can adjust the configuration
which suits them best.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>