Add support for compiling with Miri
Miri is an experimental interpreter for Rust that can check for certain
kinds of undefined behaviour.  Let's get the testsuite working with it
as much as possible.

Note that Miri currently does not support Tokio, since it doesn't
support `epoll_wait`, and it won't work with our code that calls into
Rustix (that is, anything involving the `lawn-fs` code).  We disable
these tests there to make the testsuite compile and run correctly.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>