Provide basic interfaces around credentials
It's helpful to provide some basic interfaces around credentials.  They
each have a secret, plus optionally a username, one or more URLs or URL
patterns associated with them, and additional metadata.  We also have a
form of inquiring about these credentials via a credential request.

Add various APIs around accessing these in a useful way using the
primitives provided by the protocol.

As part of this, add client and protocol handler methods that return the
ID of a message we've sent, since when we're performing certain types of
authentication, we need the ID of the last message to help chain
commands together.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>

protocol: allow sending a message and returning the ID

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>