Add a parser for the Git credential protocol
The Git credential protocol is relatively simple.  Add a parser for it
that can handle a credential request or response.

Add two possible extensions for Lawn-specific data: the authentication
type and the service that this is for.  Normally, the default service
will be `git`, but users may want to store credentials using this
protocol for other services.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>