Add a manager for various stores
Now that we have the client pieces for talking to a generic store, let's
implement a manager for a generic store and a set of traits that can be
used to implement one.  Add a parser for store paths, which can be
tricky to handle, as well as some basic tests for it.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>