Make TemplateContext owned
In the future, we'll want to provide additional data stored in this
structure, notably additional environment variables required by the
signin code for some credential helpers.  Having three separate
lifetimes is already a little silly, so let's make this data structure
own its data to avoid adding a fourth lifetime.

This will also be useful in the future for query code, where we'll want
to be able to store a context in the server for a longer period of time.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>