server: create a shared state
We'll want to have some shared state across multiple different server
connections, and in order to do that, let's create a structure for that
purpose and put an `Arc` of it into the regular server state.  In
addition, move the configuration into it, since that is shared across
multiple different connection instances.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>