template: add server context environment
We already have a client and server environment.  However, in some
cases, such as some credential backends, we need to preserve environment
variables to be passed to those backends when running commands.  This
can be useful to allow a user to log in once, get an environment
variable back, and then pass that environment variable into other calls
to that backend.  Notably, this will be useful for a future 1Password
credential helper, among others.

Let's add a server context environment which can be used to pass these
specific environment variables and which will keep them independent from
the main server environment variable state.  Add a template context
builder which can be used to easily instantiate just the required values
as well.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>