credential: keep shared state for memory backend
The memory backend's data is stored in a credential store whose data is
destroyed at latest at the end of the client session.  This is
unhelpful, because it means that every separate command-line invocation
of the Lawn client will have its own state, which makes things like
using credential helpers not work.

To fix this, let's use our special shared credential state, and convert
it to handle arbitrary data.  Then, insert our vaults into the store
once the backend is created and re-use them for subsequent sessions, so
that our data is persisted across sessions as long as the server is
running.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>