Add tools for working with script syntax
Lawn is designed to be flexible and scriptable.  To make it easy to
script commands, let's add a script syntax based off the IMAP message
syntax.

Our deserializer doesn't use Serde because Serde assumes that we can
distinctly determine the type inside of an `Option` based on context.
However, we cannot do that because all of our data is serialized as
strings, so without knowing the type of the data, it's not possible to
distinguish the hexadecimal integer `0xff` from the text or byte strings
with that same value.  This is fine for scripting, where most of the
time we're working with strings anyway, but makes decoding a bit more
difficult.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>