Implement command line interface for credentials
Add some basic command line functionality, including a script interface
and a Git credential helper.  For now, always insert the item in the
first vault in the list, but in the future, we'll allow putting it in an
arbitrary location.

As part of this, declare that we support the appropriate capability.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>