lawn: add `PATH` appropriately to tests
On FreeBSD and NetBSD, Ruby, which is used in our test fixture, is
located in a different location (`/usr/local/bin` on FreeBSD and
`/usr/pkg/bin`).  Let's change our `PATH` such that it inherits the
system one, just with our helper directory in front, so we'll always
find the correct binaries.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>