Bump MSRV to 1.63
To fix security problems, we need to update the version of
crossbeam-utils.  However, that is depended on by crossbeam-epoch, which
is depended on by flurry.  There is no way to upgrade flurry such that
crossbeam-utils is upgraded to a secure version, so we need to upgrade
to flurry 0.4.0, which depends on seize.  That crate, in turn, requires
a newer version of once_cell, which requires a newer Rust version.

Since we were going to bump the MSRV to 1.63 in July anyway in
conformance with our typical policy, and there is no other way to apply
security updates, bump the MSRV to 1.63.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>