lawn-fs: avoid clunking automatically on removal
9P requires that the remove call also effectively call clunk.  However,
not all file system backends do, and SFTP does not.  The fact that the
generic file server code did so is a legacy of it being ported over from
the original 9P code.

In addition, with flurry 0.4.0, calling clunk here causes the object
referred to by the FID to be destroyed sooner, which in turn causes the
file descriptor to be closed, which gives us EBADF.  Since we don't want
that, and the behaviour was wrong anyway, let's switch to explicitly
calling clunk where necessary to fix both problems.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>