Avoid running CI twice for both push and pull request
When we push into the same repository as the target, GitHub Actions will
kick off a job both for a push and a pull request.  This is wasteful,
because we're doing the same thing twice.

To avoid this problem, let's add a condition such that we don't run the
pull request job unless it's from a fork.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>