ssh: add handling for borrowed SSH messages
Right now, all of our messages require allocating memory.  It would be
more efficient if we could reuse one buffer for this purpose instead of
reallocating every time.  Add a BorrowedSSHMessage type that allows us
to read and write based on a buffer, and functions that can do that
reading.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>