ssh: allow reading and writing server messages at once
Split out the reading and writing paths in the SSH proxy server so that
we can both read and write requests at the same time.  Be sure to verify
that our responses happen in the same order as our requests.

Include extra logic to deal with multiple Lawn requests over the socket.
Satisfy all but the last one immediately and satisfy the last one after
a timeout to improve round-trip time.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>