ssh: set SSH socket to nonblocking when converting to Tokio
Tokio can accept Unix sockets from std, but they need to be set
nonblocking first.  However, we didn't do that before spawning the SSH
proxy or the client probe, which means that we could accidentally block
the executor.  Let's make sure that set the socket to the correct mode
to avoid this.  In the one case where we convert back to the std
variant, be sure to set it back to blocking mode.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>