ssh: substantially improve proxy performance
We've configured our server side to wait for a response if we've
received a regular message and there are no other Lawn requests
outstanding.  This means that our ping requests will be serviced
immediately.

However, we get much better performance if we lift this restriction and
always send a ping request with our non-ping request.  This means that
our throughput on a local system goes from less than 200 KiB/s to over
100 MiB/s, which is at least a 500× improvement.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>