Move socket handling to its own code
Right now, we have lots of autodetection of our socket handling in the
main code.  In the future, we'll want to provide an environment variable
which specifies the socket to which the user should connect.  To make
this easier, let's move the socket code to its own module and return a
suitable structure that can encapsulate all of the socket information.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>