Implement reading socket information from environment
Right now, we perform socket autodiscovery, which works fine in many
contexts.  However, in the future, we'll want to allow the user to
extract socket and contextual information from the environment so that
the user can script operation-specific queries in `if` shell commands.
We'll also want to allow the user to perform privileged operations on
the local machine, but not on remote machines, so include some basic
authentication information.

To do so, let's provide this data in the LAWN environment variable as a
base64-encoded CBOR blob.  This can include socket location and socket
type, but it can also include authentication information as well.
Note that when we serialize our authentication data, don't serialize
optional fields to keep the environment variable as small as possible.

Refactor out the SSH socket testing so that we can reuse it in both
cases.  In addition, add a test mode that doesn't actually perform
socket testing so that we can write some tests, and do so so that we
have more confidence in our functionality.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>