socket: improve testing of parsing environment variables
Right now, we don't support certain fields in our CBOR data, but we will
in the future.  The current parser rejects this data when parsing
because we don't know how to handle it, but just for parsing purposes,
let's add an option to parse it just for tests so we can be sure we
handle it correctly.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>