config: store template contexts in config data
In the future, we'll want people to be able to query the current
template context from the command line, such as in an `if` block.  When
we do that, we'll want to have the template context available for the
life of the operation, but we'll also want to clean it up automatically.

To make that possible, let's use a guard like for a mutex that persists
the data for the life of the operation but cleans it up when it's no
longer needed.  When we create the guard, store the template context in
the configuration struct so that we can easily access it and then drop
the data when the guard is dropped.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>