lawn: add general output format argument
We'd like people to be able to control the output format in a general
way.  Let's add an argument to pass this information in, and control how
unhandled logger errors are handled at the end of the code.

Note that we consider serialized errors to be a successful situation,
since the logging produced the expected output (even though that was an
error).

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>