socket: expose more information about socket data
To make working with this structure easier, let's expose the context
value so we can clone an existing structure and then modify the context
to create a new value.  Similarly, add a constructor to allow creating
instances more easily.

If we have a `LawnSocket` instance, allow accessing the context and the
`LawnSocketData` inside it so we can clone them.

Finally, implement a method for `LawnSocketData` to serialize it into
the value of an environment variable.  Our code already knows how to
obtain socket information from this environment variable with the
provided serialization.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>