config: allow setting socket data
In the future, we'll want to allow passing in a `LAWN` environment
variable when we run a command with a template context.  In such a case,
we'll need the socket data to serialize.  Make sure we set it any time
we look for a server socket and also whenever we spawn a server.

There are still some cases in early startup, such as when looking for
whether the system is the root machine, where the socket data won't be
set.  However, we plan to handle this gracefully in the future.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>