tests: make sure server is ready before running
When we call `run_async` on the server, we receive a blocking `File`
representing a pipe endpoint from which we can read one byte when the
server is ready to start.  Right now in our tests, we either ignore this
file descriptor and simply continue, or we read it in a blocking way,
which can block the executor and cause our tests to hang.

Let's solve this by turning this into a non-blocking file descriptor and
then actually reading from it, which guarantees that we will neither
start working with the server too early nor block the executor.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>