socket: avoid using template context before setting socket data
In general, we want to avoid using template contexts as much as possible
before setting the socket data because it creates situations where we
might accidentally fail to pass the `LAWN` environment variable to
places which need it.  In this particular case, we really only care
about the environment variables, which are accessible with another
function, so let's use that function to look them up.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>