config: generate a `LAWN` environment variable for template contexts
When we create a template context, include a `LAWN` environment variable
any time we have socket data.  If we don't log a message, so we can
detect this case if we should have had socket data, but don't fail.  We
create template contexts early on in startup to determine whether we're
the root machine and we won't have socket data yet (since we don't know
whether we should be starting a server or not), so failing would prevent
us from working at all.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>