tests: put config file in proper directory
In a future commit, we'll have tests that actually invoke the `lawn`
binary.  In order for those to work, we'll need to have the
configuration file readable by that binary, which means it needs to be
under `$HOME/.config/lawn`.  Create that directory in our tests and
store our generated configuration file in that location.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>