server: move creation of store elements to a separate thread
In a future commit, creation of store elements may call a blocking
process.  That will obviously block the executor if we do it in an async
context, so spawn a blocking thread.

This also necessarily means that the data we pass it must be at least
`Send`, if not `Sync`.  Let's update each of the objects that we pass as
`dyn Any` such that they become `dyn Any + Send + Sync + 'static`, which
allows us to move them across threads and also continue to downcast.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>