credential: allow automatically detecting credential storage location
Right now, our credential helper simply stores data in the first vault.
This may not be where the user would like it to be stored, however.
Let's make it possible for the user to configure it by providing a
configuration option, `control`, that can take an argument
(`create-location`) and produce a location as a result.

This passes the credential data in the context so that it can be
extracted and used via querying the protocol.  The Lawn client will be
updated to read and process it in a future commit.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>