main: add a tool to query contexts
When we attempt to put a credential to `/`, we'll invoke a shell to find
the location for the credential in question.  However, in many cases,
the location depends on the credential.

Add a mode to Lawn, `lawn query context`, that can query the entire
context, including credential.  Note that this requires an appropriate
`LAWN` environment variable containing the random 32-byte context ID,
which is passed only by the server to its child process and which lasts
only as long as the child process is invoked.  Thus, there is little
risk that the credential will be exposed publicly.

Mark the `context=template` protocol option as functional, which is
required for Lawn to be able to read the data.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>