Improve handling of Git credential helper
Right now, we always try to store or erase a credential from the first
vault.  That might have been valuable as an interim measure, but now
that we have the `control` configuration, we don't have to rely on that
anymore.

Instead, let's create new credentials at `/` and let the `control`
configuration tell us where to put it automatically.  For erasing, let's
iterate over each vault, and if we find the credential, erase it.

To improve the situation further, don't require a secret when parsing
credentials for erasure.  This lets users erase credentials with a
simple URL or a URL and username.

Add some rspec tests for our credential helper as well so we can be
confident it works correctly.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>