Update daemonize to 0.5.0
Version 0.4 of the daemonize crate causes some warnings from rustc about
incompatibilities in a future version of Rust.  Let's upgrade to 0.5.0
to fix that problem.

However, in 0.5.0 the `exit_action` method has been removed.  Use the
new `execute` method to get a different return for the parent and child,
and then perform the file descriptor read in the parent before exiting.
Add a comment to explain why we do this, since it may not be immediately
obvious.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>