Add a .rustfmt.toml
When using rustfmt out of the root of the repository, it's possible for
code to end up being formatted under the 2015 edition, which fails due
to the lack of async and await.  This makes automatically tidying code
in one's editor difficult.  To make this process easier, let's add a
`.rustfmt.toml` file at the root of the repository which sets the
edition correctly.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>