lawn: allow passing arguments without --
When we invoke something like `lawn run`, we often want to include
arguments.  However, older versions of clap didn't like this because it
thought that they were out-of-order arguments to the subcommand and
produced an error.

However, with newer versions of Clap, we can indeed allow these kinds of
arguments using `trailing_var_arg`, so let's do that to improve the user
experience.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>